Bug: Text overflow on overzoom.
If you zoom way too far, the text busts out of the box and makes a mess.

This uses ellipsis to beautifully indicate that there is more content in
the line, and automatically adds a scroll bar so you can scroll to see
it.

Over zoom the page to see :D